mmc: tmio: Fix tuning flow
Since tuning initialization processing was performed only for the first
time, tuning after the second time failed. Therefore, tuning
initialization processing is modified to be performed every time.

If the return value of mmc_send_tuning() is error other than -EILSEQ,
the tuning fails and process goes out of for_loop.
But the correct processing is to judge their TAP as NG.

Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>