mmc: sh_mobile_sdhi: HACK: Use suspend_late/resume_early
This patch is an interim solution to be supported suspend/resume.
If it is not applied, CPU stall has occurred with reading/writing
the SDHI resistor after resume.

Calling sequence for function:
  mmc_rescan
    mmc_power_up
      mmc_set_initial_state
        tmio_mmc_set_ios
          tmio_mmc_set_clock
            tmio_mmc_clk_stop
              sd_ctrl_read16(host, CTL_SD_CARD_CLK_CTL)
              --> CPU stall

Now root cause is not still found. However, it found that this
issue is affected to timing of suspend callback. Therefore,
it fixes to shift the suspend callback as interim solution until
root cause is found.

The modification by this patch is shown below.

Before to be applied patch:
 1) PM: suspend
        -> pm_runtime_force_suspend() in sh_mobile_sdhi
        -> tmio_mmc_host_runtime_suspend() in sh_mobile_sdhi
 2) PM: late suspend
 3) PM: noirq suspend
 4) PM: noirq resume
        -> tmio_mmc_host_runtime_resume() in sh_mobile_sdhi
 5) PM: early resume
 6) PM: resume
        -> pm_runtime_force_resume() in sh_mobile_sdhi

After to be applied patch:
 1) PM: suspend
 2) PM: late suspend
        -> pm_runtime_force_suspend() in sh_mobile_sdhi
        -> tmio_mmc_host_runtime_suspend() in sh_mobile_sdhi
 3) PM: noirq suspend
 4) PM: noirq resume
        -> tmio_mmc_host_runtime_resume() in sh_mobile_sdhi
 5) PM: early resume
        -> pm_runtime_force_resume() in sh_mobile_sdhi
 6) PM: resume

Signed-off-by: Gaku Inami <gaku.inami.xw@bp.renesas.com>