arm64: dts: r8a7795-salvator-xs: Add SATA device support
This patch adds SATA device support by default disable for Salvator-X
2nd version board on R8A7795 ES2.0 SoC.

With the Salvator-X 2nd version board, the signal of PCIE1 (CN6) and
SATA (CN8) changes with PCI Express Passive Switches. Therefore, this
devices are used by exclusion.

As for the default of this DeviceTree, PCIE1 of a board default is used.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>