v4l: vsp1: Extend VSP1 module API to allow DRM callbacks
To be able to perform page flips in DRM without flicker we need to be
able to notify the rcar-du module when the VSP has completed its
processing.

We must not have bidirectional dependencies on the two components to
maintain support for loadable modules, thus we extend the API to allow
a callback to be registered within the VSP DRM interface.

In addtion, support LIF expansion for H3(ES2.0).

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
(cherry picked from commit 6de5f2484052c18bbc5f7356ecde9498a541db5f)
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>

Conflicts:
	drivers/media/platform/vsp1/vsp1_drm.c