rcar-vin: read subdevice format for crop only when needed
Instead of caching the subdevice format each time the video device
format is set read it directly when it's needed. As it turns out the
format is only needed when figuring out the max rectangle for cropping.

This simplifies the code and makes it clearer what the source format is
used for.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Reviewed-by: Hans Verkuil <hans.verkuil@cisco.com>
(cherry picked from commit 92e3f67fb5a4fde9fcba5cf9584fd72a9c307f69)
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>