arm64: dts: r8a7796: Marks the OPP to be used during system suspend
Mark opp node to be used by cpufreq during System Suspend to RAM.
With this update, it helps to synchronize the frequency and voltage
values between software and hardware after resume.
Before this update, the frequency and voltage values
before system suspend and after system resume may be different

The opp node is selected so that the value of frequency and voltage
are as initial values of hardware after resume.

Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>