arm64: dts: r8a7796-salvator-x: Fix MSIOF clock frequency
This patch fixes the default clock frequency value of MSIOF reference
clock (MSO), and the maximum clock frequency value of MSIOF1 to the
maximum value defined as hardware specification.

The default clock frequency value of MSO updates to 133.33MHz, and the
maximum clock frequency value of MSIOF1 updates to 66.66MHz.

Fixes: 281235e3a76c ("arm64: dts: r8a7796-salvator-x: Enable MSIOF reference clock")
Fixes: d0fa39bcb81c ("arm64: dts: r8a7796-salvator-x: Enable MSIOF controller device support")
Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>