arm64: dts: r8a7796-salvator-xs: Add HSCIF1 device support
This patch adds the pin control for HSCIF1, and supports connection with
the Debug Serial-1 (CN26) for R8A7796 SoC on Salvator-X 2nd version
board.

SCIF1 and HSCI1 are sharing the pin connected to the Debug Serial-1
(CN26) on Salvator-X 2nd version board, and it is necessary to use it by
exclusion.

As for the default of this DeviceTree, SCIF1 is connected.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>