arm64: dts: r8a7796-salvator-xs: Add PCIe device support
This patch adds PCIe{0,1} device support for R8A7796 SoC on Salvator-X
2nd version board.

With the Salvator-X 2nd version board, the signal of PCIE1 (CN6) and
SATA (CN8) changes with PCI Express Passive Switches. Therefore, this
devices are used by exclusion.

As for the default of this DeviceTree, PCIE1 of a board default is used.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>