mmc: sh_mobile_sdhi: Fix SD_BUF width for R-Car Gen3 (part 2)
Commit ab0010da63c3 ("mmc: sh_mobile_sdhi: Fix SD_BUF width for R-Car
Gen3") is trying to set the SD_BUF width to 64-bit if the dedicated
DMAC is used. Unfortunately, it fails in doing so, because the
.sdbuf_64bit=true information stored in the of_rcar_gen3_compatible
structure is not properly propagated to sh_mobile_sdhi_enable_dma()
function via the `struct tmio_mmc_host *host` parameter. As a side
effect, the SD_BUF width is always set to 32bit.

Fix this.

Fixes: ab0010da63c3 ("mmc: sh_mobile_sdhi: Fix SD_BUF width for R-Car Gen3")
Signed-off-by: Eugeniu Rosca <erosca@de.adit-jv.com>
Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>