arm64: dts: r8a7796: Add placeholder for HDMI device
Add an empty device node with ports boilerplate serving as a placeholder
for the HDMI device.  HDMI is not yet supported and/or tested on R-Car
M3-W, but is supported and used on Salvator-X boards with an R-Car H3
SoC.  Hence a placeholder will allow common board support to refer to
it.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>