clk: clk-5p49x: Fix hierarchy of driver structure
The clk structure was not managed properly in this driver,
and the clk structure release was not performed normally at
the time of the remove procedure.

This patch revises the hierarchy of the driver structure and
corrects it so that the clk structure can be released properly
at the time of remove procedure.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>