media: rcar-vin: Add signal polarity select
When the type of mbus is V4L2_MBUS_BT656, the signal polarity select
register is set active high to polarity of VSYNC and HSYNC automatically.
so this patch corrects it.
In addtion, since the CHS bit in VnDMR2 register should be set when
ITU-R BT 656 mode. This patch adds its settig.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>