rcar-vin: add support for V4L2_FIELD_SEQ_{TB,BT}
The hardware do not support capturing the field types V4L2_FIELD_SEQ_TB
and V4L2_FIELD_SEQ_BT. To capture in these formats the driver needs to
change what field (top or bottom) is to be captured next and for every
other capture adjust the offset of the capture buffer.

This patch adds support for these sequential fields with the limitation
that continues capture mode is not supported. This limitation have no
visibility for user-space and the driver can already be running in
single capture mode for all other field types if the driver is not feed
buffers quickly enough.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
(cherry picked from commit d74608904e855da06f5fd3873ada25d90e828779)
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>