drm: rcar-du: add LVDS backlight support by optional GPIO
The change adds an optional possibility to set a GPIO line, which
can be used to manage panel backlight.

In general a GPIO line used to control panel's backlight should
be specified as a panel hardware property and be managed by
a gpio-backlight driver, however to preserve a compatibility with
the existing board DTS files the property is added as a DU propoerty.

Signed-off-by: Vladimir Zapolskiy <vladimir_zapolskiy@mentor.com>