pinctrl: sh-pfc: r8a77965: Add drive strength support
Define the drive strength registers for the R8A77965. Add pins which are
not part of a GPIO bank nor can be muxed between different functions but
which still allow for their drive-strength to be configured.

Based on a patch from Niklas SUderlund for the R8A7796 PFC driver.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>