arm64: dts: r8a77965-salvator-xs: Add SCIF1 support
This patch adds the pin control for SCIF1, and supports connection with
the Debug Serial-1 (CN26) on Salvator-X 2nd version board.

SCIF1 and HSCI1 are sharing the pin connected to the Debug Serial-1
(CN26) on Salvator-X 2nd version board, and it is necessary to use
it by exclusion.

As for the default of this DeviceTree, SCIF1 is connected.

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>