arm64: dts: r8a77965-salvator-x: Add CS2000 as AUDIO_CLK_B
CS2000 needs AUDIO_CLKOUT as master clock which is generated
by Renesas sound, and Renesas sound needs CS2000 as ADUIO_CLK_B.
Because of this relationship, it will be dead-lock when driver probe.

cs2000: clk_multiplier@4f {
        ...
        clocks = <&rcar_sound 0>, <&x12_clk>;
        ...
};

&rcar_sound {
        ...
        assigned-clocks = <&cs2000>;
        ...
};

This patch is using dummy audio_clkout to avoid this issue.

Based on a similar patch of the R8A7796 device tree by Kazuya Mizuguchi
<kazuya.mizuguchi.ks@renesas.com>.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>