arm64: dts: r8a77965-salvator-xs: Add SATA controller
This patch adds SATA controller support by default disable for
the Salvator-X 2nd version board on R8A77965 SoC.

With the Salvator-X 2nd version board, the signal of PCIE1 (CN6) and
SATA (CN8) changes with PCI Express Passive Switches. Therefore, this
devices are used by exclusion.

As for the default of this DeviceTree, PCIE1 of a board default is used.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>