mmc: core: Add timeout errors to retuning condition
When conditions such as temperature change, timeout may
occur due to communication failure with SD/MMC.

In case of timeout error, the re-tuning flag is not true.
Therefore, the command is retried without re-tuning.
If re-tuning is executed, the success rate of retries
can be improved.

By adding timeout to retuning conditions, re-tuning is
executed when retrying.

Signed-off-by: Masaharu Hayakawa <masaharu.hayakawa.ry@renesas.com>
[saito: rework commit message.]
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>