phy: rcar-gen3-usb2: Support System Suspend To RAM
As description in Documentation\phy.txt,
phy-core will create a device as child of phy-provider
and manage runtime pm for that device.
So runtime pm of phy-provider will be managed by
parent-child relationship. So, it needs to enable
runtime pm of phy-provider for phy-core can control
runtime pm for it when System Suspend and Resume.

Signed-off-by: Hien Dang <hien.dang.eb@renesas.com>
Signed-off-by: Junki Kato <junki.kato.xk@renesas.com>