media: rcar-vin: Fix PHTW register parameter for mbps
There is an error PHTW setting table for R8A77965 in the H/W manual.
It will be corrected to the correct value in future by version up of
H/W manual. Based on that value, this patch fixes it.

In addition, set TESTCLR bit in PHTC register when stopping csi2
in order to operate the calibration of the PHY normally.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>