media: i2c: adv748x: Fix video standard selection register setting
By video decoder user's manual, the bit 2 in Video Standard
Selection register must be reserved with the value of 1.
This driver cleared it with 0 when writing back.
Also, if this bit was not written back, a phenomenon different
from the hue captured at power on occurs.
This patch corrects it.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>