drm/bridge: adv7511: Fix i2c slave address
Problems occur when ADV7612 and ADV7511W are used at the same time.
When the slave address of the ADV7511W is specified as 0x72, HDMI related
registers will not be read on the ADV7612 side. The problem is solved by
fixing the value of the slave address of the ADV7511W.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>