iommu: ipmmu-vmsa: Add workaround for disabling uTLBs
This patch adds a workaround for disabling uTLBs on R-Car Gen3 SoCs.
Even if a master is disabled IPMMU (uTLB), the master might cause
mis-translation by issuing TLB flush from IMCTRn of a IPMMU cache
domain when IMUCTR.TTSEL value is the same with a IMCTRn's context id.
To avoid this issue, this patch just reserves context id 0 by using the
driver.

Signed-off-by: Yoshihiro Shimoda <yoshihiro.shimoda.uh@renesas.com>
Signed-off-by: Hai Nguyen Pham <hai.pham.ud@renesas.com>