soc: renesas: rcar-sysc: Update power control flow for SYSCEXTMASK register
On H3 v3.0, it can observe the system hang up after changing
power domains many times (~2000 times) continuously.
This occurs because there is conflict when changing state
of power domains.

To overcome this issue, power domain on/off control flow
need to be updated. It needs to set/clear value of SYSCEXTMASK register.

This patch is updated for above purpose.

Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Sang Nguyen <sang.nguyen.eb@renesas.com>
Signed-off-by: Tho Vu <tho.vu.wh@renesas.com>