rcar-vin: Fix IFMD table information for R8A77990
There are 8 channels in case of another VIN device, and it is a
driver that searches 0 to 7 ch at initialization and initializes it.
If this driver do not define the dummy table (excluding VIN4 and VIN5)
of IFMD at R8A77990 device, it will be initialized incorrectly as
IFMD table does not exist. This patch will fix it.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>