clk: renesas: rcar-gen3: Add support ZG clock divider for R8A77990
This patch supports the ZG clock divider for the R8A77990 SoC. In this
SoC, the ZG clock divider bits of the FRQCRB register is bit[28:24], but
this bit has a frequency setting including changing the clock source of
PLL0 and PLL1.

This patch supports only switching between 600 MHz and 300 MHz with PLL0
clock source.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>