i2c: rcar: Fix handling of the RESTART phase and STOP phase
In case of a message in read mode, setting of ICMCR.FSB or ICMCR.ESG
must be done before the last data is received.

ICMCR.FSB is set to issue stop condition.
ICMCR.FSB is included in the definition of RCAR_BUS_PHASE_STOP.

ICMCR.ESG is set to issue the restart condition.
ICMCR.ESG is included in the definition of RCAR_BUS_PHASE_START.

The setting of ICMCR.ESG must be kept from receiving the last data to
issuing the next restart condition actually on the I2C bus.

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>