media: rcar-vin: Cleanup notifier in error path
During the notifier initialization, memory for the list of associated async
subdevices is reserved during the fwnode endpoint parsing from the
v4l2-async framework. If the notifier registration fails, that memory
should be released and the notifier 'cleaned up'.

Catch the notifier registration error and perform the cleanup both for the
group and the parallel notifiers.

Link: https://patchwork.kernel.org/patch/10434533/
[koji.matsuoka.xm: the part of the patch for kernel v4.14.35 was rebased
and imported]
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>