rcar-vin: Fix setting order of UDS disable
Before stopping the capture, if UDS (Up Down Scale) is set to disable
and waiting for the capture completion, capture data invades to the area
beyond the area allocated by dma_alloc function.
Since it was obtained by the view of the H/W team that it is better to
disable UDS after stopping the capture, this patch correctes as such.
Confirmed that this problem is fixed by this correction.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>