drm: rcar-du: Add HDSR and VDSR register value check
It is prohibited to set a value less than 1 to VDS bit of VDSR
register by the H/W specification.
In addtion, it is also prohibited to set a value less than 1 to
HDS bit of HDSR register by the H/W specification.
So this patch adds value check of each registers.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>