i2c: rcar: Fix pio judgment function name
Since rc_i2c_is_dma() is a difficult-to-understand function name
from the original purpose, we modified it to a more intuitive
function name.
The aim is to determine if it is a PIO.

The reason for removing "!" Is to make it easier to judge whether
it is PIO or not.

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>