soc: renesas: rcar-sysc: 3DG Power domains for r8a7796 are always on
As in commit 6e85cefbdc08 ("soc: renesas: rcar-sysc: power domains are
always on for r8a7796"), it shows that Hardware needs to be controlled
with the state transitions do not conflict for each power domains.

After verifying again, the patterns of conflict are updated as below.

         CA5x vs CA5x      : HW exclusion control
         CA5x vs 3DG(HW)   : HW exclusion control
         CA5x vs 3DG(SW)   : SW exclusion control (Not implemented yet)
         CA5x vs Other(SW) : SW exclusion control (already implemented)
      3DG(HW) vs 3DG(HW)   : HW exclusion control
      3DG(HW) vs 3DG(SW)   : HW exclusion control
      3DG(HW) vs Other(SW) : SW exclusion control (already implemented)
      3DG(SW) vs 3DG(SW)   : SW exclusion control (already implemented)
      3DG(SW) vs Other(SW) : SW exclusion control (already implemented)
    Other(SW) vs Other(SW) : SW exclusion control (already implemented)

This patch fixes CA5x vs 3DG(SW) pattern by turning on the 3DG power
domains at always.

Signed-off-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>