arm64: dts: r8a77990-ebisu: Enable simultaneous output of VGA and HDMI
Current definition is only HDMI single output via DU0.
This patch can simultaneously output HDMI and VGA by adding definition to
VGA via DU1.

x13_clk is detached from the clocks property of the LVDS1 node.
This is because VGA does not display if x13_clkis defined in the clocks
property of the LVDS1 node when simultaneously outputting VGA and HDMI.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>