media: rcar-vin: Allow independent VIN link enablement
There is a block of code in rvin_group_link_notify() that loops through
all entities in the media graph, and prevents enabling a link to a VIN
node if any entity is in use. This prevents enabling a VIN link even if
there is an in-use entity somewhere in the graph that is independent of
the link's pipeline.

For example, the code will prevent enabling a link from the first
rcar-csi2 receiver to a VIN node even if there is an enabled link
somewhere far upstream on the second independent rcar-csi2 receiver
pipeline.

If this code is meant to prevent modifying a link if any entity in the
graph is actively involved in streaming (because modifying the CHSEL
register fields can disrupt any/all running streams), then the entities
stream counts should be checked rather than the use counts.

(There is already such a check in __media_entity_setup_link() that verifies
the stream_count of the link's source and sink entities are both zero,
but that is insufficient, since there should be no running streams in
the entire graph).

Modify the media_device_for_each_entity() loop to check the entity
stream_count instead of the use_count, and elaborate on the comment.
VIN node links can now be enabled even if there are other independent
in-use entities that are not streaming.

Fixes: c0cc5aef31 ("media: rcar-vin: add link notify for Gen3")

Signed-off-by: Steve Longerbeam <slongerbeam@gmail.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>
Link: https://patchwork.kernel.org/patch/10749655/
[koji.matsuoka.xm: the part of the patch for kernel v4.14.75
was rebased and imported]
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>