drm: rcar-du: Fix loop procedure in scatter gather page set
Originally the variable of "i" is used for the loop count of
the number of format planes in rcar_du_vsp_plane_prepare_fb
function and should be not used the same "i" as the loop count
of sg_set_page.

It is affected when the number of format planes is 2 (NV12 format
etc.) and 3 (YUV420 format etc.).
For the number of format planes is 1 (RGB565, ARGB888, YUYV etc),
this function works fine.

This patch solves its problem.

Fixes: c424cd4724 ("drm: rcar-du: Allow importing non-contiguous
dma-buf with VSP")

Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>