phy: rcar-gen3-usb2: Correct VBUS behavior at over-current
When VBCTRL.OCCLREN is set, VBCTRL.VBOUT and ADPCTRL.DRVVBUS
are automatically cleared when an over-current occurs.
On OTG channels, clearing these bits will turn off VBUS
so device will appear to be disconnected on the bus.
Clearing VBCTRL.OCCLREN will resolve this issue by retaining VBUS status.
USB port is able to be recovered automatically and the device can be
connected again.

Signed-off-by: Kazuya Mizuguchi <kazuya.mizuguchi.ks@renesas.com>
Signed-off-by: Tho Vu <tho.vu.wh@renesas.com>